requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==25.2.0
idna==3.7
inflection==0.3.1
iniconfig==2.0.0
lazy-object-proxy==1.10.0
oauthlib==3.2.2
packaging==24.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.1
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
